// server/src/types/Catalog.ts
export interface ArqueoDetalle {
  codigo: number;
  aperturaCodigo: number;
  fecha: string;
  descripcion: string | null;
  tipo: string;
  monedaCodigo: number | null;
  monedaDenominacion: string | null;
  monedaSimbolo: string | null;
  tasa: number | null;
  cantidad: number | null;
  monto: number;
}

export interface ArqueoDetalleInput {
  monedaCodigo: number;
  cantidad: number;
}

export interface ArqueoCaja {
  codigo: number;
  aperturaCodigo: number;
  fecha: string;
  estado: number;
  cajaCodigo: number;
  cajaDescripcion: string;
  usuarioCodigo: number;
  usuarioUsername: string;
  usuarioNombre: string;
  usuarioApellido: string;
  montoApertura: number;
  saldoAnterior: number;
  totalCreditos: number;
  totalDebitos: number;
  total: number;
  motivo: string | null;
  detalles: ArqueoDetalle[];
}